// sourcecode/BTrak.Packages/snova-billing-module/billing-module-components/billing-components/src/lib/billing/models/site-model.ts
export class SiteModel {
  name: string;
  email: string;
  address: string;
  addressee: string;
  companyId: string;
  createdDatetime: any;
  createdByUserId: string;
  searchText: string;
  isArchived: boolean;
  timeStamp: any;
  id: string;
  siteId: string;
  autoCTariff: any;
  roofRentalAddress: string;
  parcellNo: string;
  muncipallity: string;
  canton: string;
  date: Date;
  m2:any;
  chf:any;
  term:Number;
  startingYear: Date;
  productionFirstYear:Number;
  autoCExpected:any;
  annualReduction:any;
  repriceExpected:any;
}